# pkg/trace: avoid sorting custom aggregation tags in assembleGrain