--- a/releasenotes/notes/trace-agent-no-sorting-on-custom-aggregations-6e82a90fe47508b8.yaml
+++ b/releasenotes/notes/trace-agent-no-sorting-on-custom-aggregations-6e82a90fe47508b8.yaml
@@ -1,0 +1,11 @@
+# Each section from every releasenote are combined when the
+# CHANGELOG.rst is rendered. So the text needs to be worded so that
+# it does not depend on any information only available in another
+# section. This may mean repeating some details, but each section
+# must be readable independently of the other.
+#
+# Each section note must be formatted as reStructuredText.
+---
+enhancements:
+  - |
+    APM: improved stats computation by removing not needed sorting step.