--- a/releasenotes/notes/trace-agent-no-sorting-on-custom-aggregations-6e82a90fe47508b8.yaml
+++ b/releasenotes/notes/trace-agent-no-sorting-on-custom-aggregations-6e82a90fe47508b8.yaml
@@ -8,4 +8,4 @@
 ---
 enhancements:
   - |
-    APM: improved stats computation by removing not needed sorting step.
+    APM: improved stats computation by removing unneeded sorting step.