### What does this PR do?

This PR leverages the fact that custom aggregations are already sorted (cf [here](https://github.com/DataDog/datadog-agent/blob/7ca5a98077e0f16e922a0e0bb223b862f5d1f6e2/pkg/trace/stats/concentrator.go#L68)) to avoid the sort operation on all insertions with custom aggregations and remove some allocations

**Benchmark**

Before:
BenchmarkHandleSpanRandom-4                 	  100899	     11231 ns/op	    5641 B/op	      50 allocs/op
BenchmarkHandleSpanRandomWithCustomAggr-4   	   59293	     21202 ns/op	   12620 B/op	      86 allocs/op

After:
BenchmarkHandleSpanRandom-4                 	   93445	     11684 ns/op	    5642 B/op	      50 allocs/op
BenchmarkHandleSpanRandomWithCustomAggr-4   	   79399	     15239 ns/op	   10954 B/op	      62 allocs/op

### Motivation

What inspired you to submit this pull request?

### Additional Notes

Anything else we should know when reviewing?

### Describe your test plan

Write there any instructions and details you may have to test your PR.