@bmermet 2020-12-08
https://github.com/DataDog/datadog-agent/pull/6842 made this PR obsolete.